Changed randomness seed to circuit1 test#42

+++ include/nil/crypto3/marshalling/zk/types/placeholder/proof.hpp
@@ -113,10 +113,11 @@ namespace nil {
                     std::tuple<
                         // typename commitment_scheme_type::commitment_type commitments
                         // All of them are similar size. So no nil::marshalling::option needed
+                        // There won't be too many commitments. So uint8_t is enough for length
                         nil::marshalling::types::array_list<
                             TTypeBase,
                             typename commitment<TTypeBase, typename Proof::commitment_scheme_type>::type,
-                            nil::marshalling::option::sequence_size_field_prefix<nil::marshalling::types::integral<TTypeBase, std::size_t>>
+                            nil::marshalling::option::sequence_size_field_prefix<nil::marshalling::types::integral<TTypeBase, std::uint8_t>>
                         >,
 
                         // evaluation_proof eval_proof
@@ -133,7 +134,7 @@ namespace nil {
                     nil::marshalling::types::array_list<
                         TTypeBase,
                         typename commitment<TTypeBase, typename Proof::commitment_scheme_type>::type,
-                        nil::marshalling::option::sequence_size_field_prefix<nil::marshalling::types::integral<TTypeBase, std::size_t>>
+                        nil::marshalling::option::sequence_size_field_prefix<nil::marshalling::types::integral<TTypeBase, std::uint8_t>>
                     > filled_commitments;
                     for( const auto &it:proof.commitments){
                         filled_commitments.value().push_back(
